+++ nutanix/hosts.go
@@ -1,4 +1,3 @@
-
 package nutanix
 
 import (
@@ -6,23 +5,22 @@ import (
 )
 
 type HostResponse struct {
-	Metadata	*HostMetadata
-	Entities	[]HostEntity
+	Metadata *HostMetadata
+	Entities []HostEntity
 }
 
 type HostMetadata struct {
-
 }
 
 type HostEntity struct {
-	Name		string
-	CpuFrequency	float64	`json:"cpu_frequency_in_hz"`
-	CpuCapacity	float64	`json:"cpu_capacity_in_hz"`
-	MemoryCapacity	float64	`json:"memory_capacity_in_bytes"`
-	NumVms		int	`json:"num_vms"`
-	BootTime	float64	`json:"boot_time_in_usecs"`
-	Stats		map[string]string
-	UsageStats	map[string]string
+	Name           string
+	CpuFrequency   float64 `json:"cpu_frequency_in_hz"`
+	CpuCapacity    int     `json:"cpu_capacity_in_hz"`
+	MemoryCapacity int     `json:"memory_capacity_in_bytes"`
+	NumVms         int     `json:"num_vms"`
+	BootTime       float64 `json:"boot_time_in_usecs"`
+	Stats          map[string]string
+	UsageStats     map[string]string
 }
 
 func (n *Nutanix) GetHosts() []HostEntity {
